add darwin_arm64 example providers lock command
Updates the list of ubiquitous platforms in the `providers lock` command to include the now common ARM macOS machines.

We've been seeing lock files being modified in unrelated (that is, not updating a provider) terraform changes because the author of the change is the first to run it on an M1 machine. When I grabbed the list of platforms from this page to update the configuration, I noticed I wasn't seeing the changes I expected because `darwin_arm64` wasn't on the list.